refactored update, removed redundant CouponMap

+++ sketches/src/test/java/com/yahoo/sketches/hll/HllMapTest.java
@@ -37,7 +37,7 @@ public class HllMapTest {
     key = intToBytes(1, key);
     for (int i=1; i<= u; i++) {
       id = intToBytes(i, id);
-      int coupon = Map.coupon16(id);
+      short coupon = (short) Map.coupon16(id);
       est = map.update(key, coupon);
       if (i % 100 == 0) {
         double err = (est/i -1.0) * 100;
@@ -79,7 +79,7 @@ public class HllMapTest {
       key = intToBytes(j, key);
       for (i=0; i< u; i++) {
         id = longToBytes(++v, id);
-        int coupon = Map.coupon16(id);
+        short coupon = (short) Map.coupon16(id);
         map.update(key, coupon);
       }
       double est = map.getEstimate(key);
